NEW: Add optional phpstan execution to PHPCS_TEST
PHPStan is a static analyser that may be of use to some repositories. Notably, I've got a small, passing PHPStan level 1 PR to submit to framework.

This will include a phpstan call in the PHPCS_TEST build, if phpstan.neon.dist exists.

I've opted not to split it out as a separate build var to reduce the number of separate builds. This is essentially another form of sophisticated listing, hence putting it there,